Fixed printing, still need to find a way to make sure that any less-than signs are printed properly
If a less-than sign is found when Fuchsia repeats a portion of your
speech, it’s printed, meaning that this could allow the user to print
HTML tags which should not be allowed.

I could possibly add a function that opens windows/tabs that parses the
text back for a Google search or opening of webpage.